requirements.txt
Django==3.1.1
djangorestframework==3.12.1

# Code quality
black==20.8b1
mypy==0.782
pre-commit==2.7.1
isort==5.5.3

# test
coverage==5.3